+++ boot/src/test/java/io/github/tgkit/boot/TelegramBotRunnerTest.java
@@ -17,8 +17,8 @@ package io.github.tgkit.boot;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import io.github.tgkit.internal.bot.Bot;
-import io.github.tgkit.internal.init.BotCoreInitializer;
+import io.github.tgkit.api.bot.Bot;
+import io.github.tgkit.api.init.BotCoreInitializer;
 import io.github.tgkit.testkit.TelegramMockServer;
 import java.util.List;
 import java.util.concurrent.TimeUnit;
